+++ app/src/main/java/com/rama/mako/managers/PrefsManager.kt
@@ -58,8 +58,10 @@ class PrefsManager private constructor(context: Context) {
         const val APPS_ICON_PACK_PACKAGE = "apps:icon_pack_package"
         const val APPS_PROFILE_INDICATOR = "apps:profile_indicator"
 
-        //        const val APPS_SHOW_HIDDEN = "apps:show_hidden"
         const val HOME_BACKGROUND_MODE = "home:background_mode"
+        const val HOME_DOUBLE_TAP_SLEEP = "home:double_tap_sleep"
+        const val HOME_BACKGROUND_MODE_SCREEN_OPACITY_STRENGTH =
+            "home:background_mode:screen_opacity_strength"
         const val GROUPS_IDS = "groups:ids"
         const val GROUPS_HEADERS = "groups:headers"
         const val GROUPS_COLLAPSIBLE = "groups:collapsible"
@@ -71,11 +73,13 @@ class PrefsManager private constructor(context: Context) {
         const val BATTERY_CHARGE_STATUS = "battery:charge_status"
         const val CLOCK_FORMAT = "clock:format"
         const val CLOCK_APP = "clock:app"
+        const val DATE_APP = "date:app"
         const val FONT_STYLE = "font:style"
         const val FONT_CUSTOM_PATH = "font:custom_path"
         const val APP_LANGUAGE = "app:language"
         const val MIGRATION_ICON_SOURCE_RADIO = "migration:icon_source_radio"
         const val SYSTEM_BAR_VISIBLE = "system:bar:visible"
+        const val SYSTEM_PREVENT_ROTATION = "system:prevent_rotation"
 
         const val APP_THEME_NAME = "app:theme:name"
         const val APP_THEME_H1 = "app:theme:clock"
@@ -173,6 +177,7 @@ class PrefsManager private constructor(context: Context) {
         const val RAMA = "rama"
         const val MAKO = "mako"
         const val CATPPUCCIN_MOCHA = "catppuccin_mocha"
+        const val CATPPUCCIN_LATTE = "catppuccin_latte"
         const val DRACULA = "dracula"
         const val MELANGE = "melange"
         const val TOKYO_NIGHT = "tokyo_night"
@@ -205,6 +210,7 @@ class PrefsManager private constructor(context: Context) {
 
                 .putString(PrefKeys.CLOCK_FORMAT, ClockFormat.HOUR_24)
                 .putString(PrefKeys.CLOCK_APP, "")
+                .putString(PrefKeys.DATE_APP, "")
                 .putString(PrefKeys.APP_LANGUAGE, Language.SYSTEM)
 
                 .putFloat(PrefKeys.APP_UI_SCALE, 1f)
@@ -215,7 +221,10 @@ class PrefsManager private constructor(context: Context) {
                 .putString(PrefKeys.APPS_ICON_SOURCE, IconSource.NONE)
                 .putString(PrefKeys.APPS_ICON_PACK_PACKAGE, "")
                 .putString(PrefKeys.HOME_BACKGROUND_MODE, BackgroundMode.DEFAULT)
+                .putBoolean(PrefKeys.HOME_DOUBLE_TAP_SLEEP, false)
+                .putInt(PrefKeys.HOME_BACKGROUND_MODE_SCREEN_OPACITY_STRENGTH, 9)
                 .putBoolean(PrefKeys.SYSTEM_BAR_VISIBLE, false)
+                .putBoolean(PrefKeys.SYSTEM_PREVENT_ROTATION, false)
 
                 .putString(PrefKeys.APP_THEME_NAME, Theme.DRACULA)
 
@@ -235,6 +244,7 @@ class PrefsManager private constructor(context: Context) {
 
                 .putBoolean(PrefKeys.SETTINGS_SECTION_CLOCK, true)
                 .putBoolean(PrefKeys.SETTINGS_SECTION_TEMPERATURE, true)
+
                 .putBoolean(PrefKeys.SETTINGS_SECTION_BACKGROUND, true)
                 .putBoolean(PrefKeys.SETTINGS_SECTION_DATE, true)
                 .putBoolean(PrefKeys.SETTINGS_SECTION_BATTERY, true)
@@ -400,6 +410,12 @@ class PrefsManager private constructor(context: Context) {
     fun isSystemBarVisible(): Boolean =
         prefs.getBoolean(PrefKeys.SYSTEM_BAR_VISIBLE, false)
 
+    fun isDoubleTapToSleepEnabled(): Boolean =
+        prefs.getBoolean(PrefKeys.HOME_DOUBLE_TAP_SLEEP, false)
+
+    fun setDoubleTapToSleepEnabled(enabled: Boolean) =
+        prefs.edit().putBoolean(PrefKeys.HOME_DOUBLE_TAP_SLEEP, enabled).apply()
+
     // SETTINGS - CLOCK
 
     fun getClockFormat(): String =
@@ -413,6 +429,12 @@ class PrefsManager private constructor(context: Context) {
 
     fun setClockApp(appId: String) =
         prefs.edit().putString(PrefKeys.CLOCK_APP, appId).apply()
+
+    fun getDateApp(): String =
+        prefs.getString(PrefKeys.DATE_APP, "") ?: ""
+
+    fun setDateApp(appId: String) =
+        prefs.edit().putString(PrefKeys.DATE_APP, appId).apply()
 
     // SETTINGS - DATE
 
@@ -453,6 +475,14 @@ class PrefsManager private constructor(context: Context) {
     // SETTINGS - FONT
 
     // SETTINGS - BACKGROUND
+
+    fun getHomeBackgroundScreenOpacityStrength(): Int {
+        return prefs.getInt(PrefKeys.HOME_BACKGROUND_MODE_SCREEN_OPACITY_STRENGTH, 9)
+    }
+
+    fun setHomeBackgroundScreenOpacityStrength(strength: Int) {
+        prefs.edit().putInt(PrefKeys.HOME_BACKGROUND_MODE_SCREEN_OPACITY_STRENGTH, strength).apply()
+    }
 
     fun getHomeBackgroundMode(): String {
         return when (prefs.getString(PrefKeys.HOME_BACKGROUND_MODE, BackgroundMode.DEFAULT)) {