+++ changelog.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## Backlog
+- Add gesture section
+- Add option for home double tap
+- Add option for home long press (currently shows settings)
+- Move clock tap option to gestures
+- Move date tap option to gestures
+- Add option for clock long press
+- Add option for date long press
+- Add option for sleep among the apps that can be selected for gestures
+- Add option for open settings among the apps that can be selected for gestures
+
+## 41
+- Screen opacity control
+- Fix navigation bar color + opacity
+- Add catppuccin latte light theme
+- Prevent screen rotation checkbox
+- Add clouds wallpaper
+- Add mountains wallpaper
+- Add pond wallpaper
+- Add double tap to put phone to sleep
+- Add date tap action
+- Replace context menu with menu bar
+- Add multiple selection option
+
 ## 40
 - Add zoom control
 - Prevent non letter for been the initial of the profiles